+++ examples/simple/search_indexes/viewsets/__init__.py
@@ -1,6 +1,6 @@
 from .address import AddressDocumentViewSet
 from .author import AuthorDocumentViewSet
-from .book import BookDocumentViewSet
+from .book import BookDocumentViewSet, BookOrderingByScoreDocumentViewSet
 from .city import CityDocumentViewSet
 from .publisher import PublisherDocumentViewSet
 
@@ -8,6 +8,7 @@ __all__ = (
     'AddressDocumentViewSet',
     'AuthorDocumentViewSet',
     'BookDocumentViewSet',
+    'BookOrderingByScoreDocumentViewSet',
     'CityDocumentViewSet',
     'PublisherDocumentViewSet',
 )